Handler.Feature propagate mousedown/touchstarts.
The previous commits fix the issue where a touchstart was causing the
feature handler to fire, which then could confuse other handlers that
thought the touch was still happening (leading to multi-touch events,
e.g. the map zooming when you tried to pan it). However, dragging the
map starting on a feature then did nothing. We flip the flag to start
propagating these events, which allows dragging on features to work.